backend/requirements.txt
streamlit
python-dotenv
pyngrok
fastapi
nest-asyncio
pydantic
reportlab
md2pdf
cloudinary
pdfminer2
flask==2.2.3
flask_restful
flask_mongoengine
flask_jwt_extended
flask_cors
englisttohindi
llama_index
sentence-transformers
w3storage
